db_readwrite/load_generator/requirements.txt
psycopg[binary]==3.3.0
psycopg[pool]==3.3.0
prometheus-client==0.24.1
pyyaml==6.0.1
